+++ submissions/myproject/package.json
@@ -5,30 +5,30 @@
     "main": "src/index.js",
     "type": "module",
     "scripts": {
-      "start": "node src/index.js",
-      "dev": "nodemon src/index.js",
-      "test": "echo \"Error: no test specified\" && exit 1"
+        "start": "node src/index.js",
+        "dev": "nodemon src/index.js",
+        "test": "echo \"Error: no test specified\" && exit 1"
     },
     "keywords": [
-      "agent",
-      "twitter",
-      "trends",
-      "subnet59",
-      "mcp",
-      "ai"
+        "agent",
+        "twitter",
+        "trends",
+        "subnet59",
+        "mcp",
+        "ai"
     ],
     "author": "Naesmal",
     "license": "MIT",
     "dependencies": {
-      "@modelcontextprotocol/sdk": "^1.10.1",
-      "agent-twitter-client": "^0.0.18",
-      "dotenv": "^16.3.1",
-      "node-cron": "^3.0.3",
-      "stopword": "^2.0.8",
-      "winston": "^3.11.0"
+        "@modelcontextprotocol/sdk": "^1.10.1",
+        "agent-twitter-client": "^0.0.18",
+        "dotenv": "^16.3.1",
+        "node-cron": "^3.0.3",
+        "openai": "^4.95.1",
+        "stopword": "^2.0.8",
+        "winston": "^3.11.0"
     },
     "devDependencies": {
-      "nodemon": "^3.0.2"
+        "nodemon": "^3.0.2"
     }
-  }
-  
+}
